Update major.minor.patch version for 64bit ints
Update NewVersion to parse ints with a 64bit int size.

Version's major, minor and patch fields are all int64 but NewVersion was
parsing them with a 32bit bitsize, update this to 64bits.

This is useful when the patch number is a timestamp in the format:
yyMMddHHmmssSSS

Add a test cases using 2147483648 (1 more than max 32bit int).