Constraints are brought in line with other tools
* >11 now is looking for 12.0.0 and up rather than >11.0.0
  This is more intuitive and what other tools like npm do. This
  applies to all segments. >1.2 is now looking for 1.3.0 and up
  instead of 1.2.0 and up.
* Validation for = or "" is always the same message. It is a
  basic message even when ranges occur.

Note, nodejs semver does not follow strict semver on prereleases.
That semver package for a range of >1.2.3-alpha.1 would not find
2.3.4-beta.2 to be in the range. Prerelease search is scoped to
the version rather than all versions. In their notes they say
this is not technically semver to do if you follow the precidence
rules. This package follows the semver precidence rules.