Update range for ^ to follow npm/js and rust
Previously the ^ syntax below major version 0 round up to 1. Now
0.x.y and 0.x.z are considered compatible when y >= z. This is how
rust and npm/js operate.

The exception to this is when the minor is below 0. At that point
every release is considered not compatible with anything else.
That means 0.0.x and 0.0.y are not considered compatible even when
x > y.

Previously the PHP/composer pattern had been followed.

In practice some tools, like Helm charts, had followed the
npm/rust method for versioning. This bring the ^ in line with
real world usage desires for this library.