Limit prerelease detection to individual ranges
Previously, I was detecting if any range in a larger constraint set contained a prerelease, and allowed prereleases for any constraint in the set. This incorrectly allowed prereleases on constraints that did not participate in a range that had prereleases.

When rewriteRange is called, we lose key information about if a constraint was originally part of a range:

1.0.0-alpha.1 - 1.0.0 becomes >=1.0.0-alpha.1, <=1.0.0

I have updated rewriteRange so that it also returns any constraints that were originally part of a range that used a prerelease. Then when we are parsing constraints, allowPrerelease on that constraint is enabled if it was originally detected as participating in a range with prereleases. This ensures that only the individual ranges have allowPrerelease enabled.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>