Fix for allowing some version that were invalid
The NewVersion function, which uses the loose parser, had the regex for
detection updated based on the official one. A change was made to allow
for versions like 1.2 and other "loose" ones (to use the node semver
term).

StrictNewVersion had some internal validation updated to catch issues.

For NewVersion, the benchmarking is now faster than the previous regex.

Ref #211

Signed-off-by: Matt Farina <matt.farina@suse.com>