Handle pre-releases on all in an and group
Up to this point, all versions in an and group needed to have a
prerelease to be checked with those correctly. This change
enables just one item in an and group to have a pre-release
and engage it for all items in that and group.

This does not work across ORs. For example,
`>= 1.2.3 < 2.0.0-beta || > 3` will evaluate prereleases in the
group between 1.2.3 and 2.0.0-beta but will not look at
prereleases when looking at `>3`.

Signed-off-by: Matt Farina <matt.farina@suse.com>